requirements.txt
fastapi
uvicorn[standard]
playwright
Pillow
python-multipart
